+++ CHANGELOG.md
@@ -32,6 +32,7 @@ for now removed features.
 - **com-nav-item**: now the focus ring is visible on the com-nav-item when it is focused (regression introduce with design token)
 - **galactic-menu**: accessibility issue with the contrast of overlay
 - **counter**: fix focus style for Firefox
+- **form-field**: move hint and error message in aria-label instead of put in aria-description
 
 ### Security